Parallel flow add join operation

+++ src/main/java/com/deep/crow/task/parallel/MultiOrder.java
@@ -63,7 +63,7 @@ class MultiOrder<T> implements Comparable<MultiOrder<T>> {
         if (Objects.isNull(o)) {
             return -1;
         }
-        return (int) (order - o.order);
+        return order - o.order;
     }
 
     @Override
